Code-Bereinigungen in Modul `ReducedInput`;
