+++ ReducedInput/src/de/elamx/reducedinput/dataobjects/LayerData.java
@@ -1,0 +1,57 @@
+/*
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
+ */
+package de.elamx.reducedinput.dataobjects;
+
+import de.elamx.laminate.failure.Criterion;
+
+/**
+ *
+ * @author Florian Dexl
+ */
+public class LayerData extends ReducedInputDataObject {
+
+    private Double thickness;
+    private Double angle;
+    private String materialName;
+
+    private Criterion criterion = null;
+
+    public LayerData(String name) {
+        super(name);
+    }
+
+    public Double getThickness() {
+        return thickness;
+    }
+
+    public void setThickness(Double thickness) {
+        this.thickness = thickness;
+    }
+
+    public Double getAngle() {
+        return angle;
+    }
+
+    public void setAngle(Double angle) {
+        this.angle = angle;
+    }
+
+    public String getMaterialName() {
+        return materialName;
+    }
+
+    public void setMaterialName(String materialName) {
+        this.materialName = materialName;
+    }
+
+    public Criterion getCriterion() {
+        return criterion;
+    }
+
+    public void setCriterion(Criterion criterion) {
+        this.criterion = criterion;
+    }
+
+}
