+++ ReducedInput/src/de/elamx/reducedinput/dataobjects/ReducedInputDataObject.java
@@ -1,0 +1,26 @@
+/*
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
+ */
+package de.elamx.reducedinput.dataobjects;
+
+/**
+ *
+ * @author Florian Dexl
+ */
+public abstract class ReducedInputDataObject {
+
+    private String name;
+
+    public ReducedInputDataObject(String name) {
+        this.name = name;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+}
